Move IIndirectReferenceResolver to Services
Matching directory to namespace
