--- SafeRapidPdf/ObjectResolver/IIndirectReferenceResolver.cs
@@ -1,11 +1,0 @@
-﻿using SafeRapidPdf.Objects;
-
-namespace SafeRapidPdf
-{
-    public interface IIndirectReferenceResolver
-    {
-        PdfXRef XRef { get; }
-
-        PdfIndirectObject GetObject(int objectNumber, int generationNumber);
-    }
-}
