+++ SafeRapidPdf/Services/IIndirectReferenceResolver.cs
@@ -1,0 +1,11 @@
+﻿using SafeRapidPdf.Objects;
+
+namespace SafeRapidPdf.Services
+{
+    public interface IIndirectReferenceResolver
+    {
+        PdfXRef XRef { get; }
+
+        PdfIndirectObject GetObject(int objectNumber, int generationNumber);
+    }
+}
